tag-manager: Erase starred URIs on trash & delete
If a starred item was deleted or trashed, it's no longer displayed in
the Starred view, which ignores broken URIs in the database. Erasing the
broken URI in the database is not necessary for this.

However, if we keep the broken URI in the database and later a different
file appears at the same URI, it's going to be treated as if it starred
was starred, which is wrong.

In order to avoid this, delete URIs from the database when the item they
point to is deleted/trashed by this application.

This has an undesirable side effect: when restoring previously starred
items from the trash, they don't get their star restored. This needs to
be addressed by the undo machinery.