undo-operations: Restore stars if trashing is undone
Undo operations should aim to restore the previous state as much as
possible.

Thus, if starred files were moved to trash, let's store that information
in order to get them starred again if this operation is undone.

Note: This doesn't apply to the "Restore from Trash" action performed
directly in trash:/// contents.